requirements.txt
asgiref==3.10.0
Django==5.2.7
django-filter==25.2
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
sqlparse==0.5.3
whitenoise==6.11.0
Pillow>=10.0.0
dj-database-url==2.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
